CHANGE: moved fetch_job() to utility_functions.py
