+++ app/routes/utility_functions.py
@@ -3,10 +3,12 @@ import tarfile
 import zipfile
 import json
 import logging
+import redis
 from modules.loggingFunctions import initialize_logging
 from flask import current_app
 from werkzeug.utils import secure_filename
 from rdflib import URIRef
+from rq import Queue
 
 # logging
 log_file = initialize_logging()
@@ -59,13 +61,6 @@ def fix_uri(s):
     uri = URIRef(s)
     return uri
 
-def is_json(myjson):
-  try:
-    json_object = json.loads(myjson)
-  except:
-    return False
-  return True
-
 def to_readable(values,readable):
     '''
     Converts URI to human readable form.
@@ -93,3 +88,17 @@ def to_readable(values,readable):
         except:
             log.error('to_readable(): No readable form found for ' + values)
             return values
+
+def fetch_job(job_id):
+    '''
+    Iterates through all queues looking for the job.
+    '''
+    redis_url = current_app.config['REDIS_URL']
+    redis_connection = redis.from_url(redis_url)
+    queues = current_app.config['QUEUES_SPFY']
+    for queue in queues:
+        q = Queue(queue, connection=redis_connection)
+        job = q.fetch_job(job_id)
+        if job is not None:
+            return job
+    return {'is_failed': True, 'exc_info': 'job not found'}
